+++ src/ripple_detection/detectors/_hse.py
@@ -4,6 +4,7 @@ import numpy as np
 import pandas as pd
 from numpy.typing import ArrayLike
 
+from ripple_detection._call_hints import explain_call_errors
 from ripple_detection.core import (
     _is_immobile_at_endpoints,
     get_multiunit_population_firing_rate,
@@ -29,6 +30,7 @@ from ripple_detection.detectors._validation import (
 )
 
 
+@explain_call_errors
 def multiunit_HSE_detector(
     time: ArrayLike,
     multiunit: ArrayLike,